マイグレーションファイル修正、既存seedファイルをDatabaseSeederに追加

+++ database/migrations/2018_03_15_145738_mod_saved_designs_table_add_column_upload_files.php
@@ -25,8 +25,10 @@ class ModSavedDesignsTableAddColumnUploadFiles extends Migration
      */
     public function down()
     {
-        Schema::table('saved_designs', function (Blueprint $table) {
-            $table->dropColumn('uploaded_files');
-        });
+        if (Schema::hasTable('saved_designs')) {
+            Schema::table('saved_designs', function (Blueprint $table) {
+                $table->dropColumn('uploaded_files');
+            });
+        }
     }
 }
